requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
discord==1.0.1
discord.py==1.5.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.5.1
